// src/generated/pistonpanel/common.ts
export interface StringSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: string def = 3;
     */
    def: string;
    /**
     * @generated from protobuf field: pistonpanel.v1.StringSetting.InputType input_type = 10;
     */
    inputType: StringSetting_InputType;
    /**
     * @generated from protobuf field: string placeholder = 6;
     */
    placeholder: string;
    /**
     * @generated from protobuf field: int32 min_length = 7;
     */
    minLength: number;
    /**
     * @generated from protobuf field: int32 max_length = 8;
     */
    maxLength: number;
    /**
     * @generated from protobuf field: string pattern = 9;
     */
    pattern: string;
    /**
     * @generated from protobuf field: bool disabled = 11;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf enum pistonpanel.v1.StringSetting.InputType
 */
export enum StringSetting_InputType {
    /**
     * @generated from protobuf enum value: TEXT = 0;
     */
    TEXT = 0,
    /**
     * @generated from protobuf enum value: PASSWORD = 1;
     */
    PASSWORD = 1,
    /**
     * @generated from protobuf enum value: EMAIL = 2;
     */
    EMAIL = 2,
    /**
     * @generated from protobuf enum value: SEARCH = 3;
     */
    SEARCH = 3,
    /**
     * @generated from protobuf enum value: TEL = 4;
     */
    TEL = 4,
    /**
     * @generated from protobuf enum value: URL = 5;
     */
    URL = 5,
    /**
     * @generated from protobuf enum value: TEXTAREA = 6;
     */
    TEXTAREA = 6
}
/**
 * @generated from protobuf message pistonpanel.v1.IntSetting
 */
export interface IntSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: int32 def = 3;
     */
    def: number;
    /**
     * @generated from protobuf field: int32 min = 4;
     */
    min: number;
    /**
     * @generated from protobuf field: int32 max = 5;
     */
    max: number;
    /**
     * @generated from protobuf field: int32 step = 6;
     */
    step: number;
    /**
     * @generated from protobuf field: string placeholder = 7;
     */
    placeholder: string;
    /**
     * @generated from protobuf field: bool thousand_separator = 8;
     */
    thousandSeparator: boolean;
    /**
     * @generated from protobuf field: bool disabled = 9;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.DoubleSetting
 */
export interface DoubleSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: double def = 3;
     */
    def: number;
    /**
     * @generated from protobuf field: double min = 4;
     */
    min: number;
    /**
     * @generated from protobuf field: double max = 5;
     */
    max: number;
    /**
     * @generated from protobuf field: double step = 6;
     */
    step: number;
    /**
     * @generated from protobuf field: string placeholder = 7;
     */
    placeholder: string;
    /**
     * @generated from protobuf field: bool thousand_separator = 8;
     */
    thousandSeparator: boolean;
    /**
     * @generated from protobuf field: int32 decimal_scale = 9;
     */
    decimalScale: number;
    /**
     * @generated from protobuf field: bool fixed_decimal_scale = 10;
     */
    fixedDecimalScale: boolean;
    /**
     * @generated from protobuf field: bool disabled = 11;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.BoolSetting
 */
export interface BoolSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: bool def = 3;
     */
    def: boolean;
    /**
     * @generated from protobuf field: bool disabled = 4;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.ComboSetting
 */
export interface ComboSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * List of options
     *
     * @generated from protobuf field: repeated pistonpanel.v1.ComboSetting.Option options = 3;
     */
    options: ComboSetting_Option[];
    /**
     * @generated from protobuf field: string def = 4;
     */
    def: string;
    /**
     * @generated from protobuf field: bool disabled = 5;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.ComboSetting.Option
 */
export interface ComboSetting_Option {
    /**
     * Sent to server
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * Displayed to user
     *
     * @generated from protobuf field: string display_name = 2;
     */
    displayName: string;
    /**
     * May be used for an icon for an option
     *
     * @generated from protobuf field: optional string icon_id = 3;
     */
    iconId?: string;
    /**
     * keywords to also use for filtering
     *
     * @generated from protobuf field: repeated string keywords = 4;
     */
    keywords: string[];
}
/**
 * @generated from protobuf message pistonpanel.v1.StringListSetting
 */
export interface StringListSetting {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: repeated string def = 3;
     */
    def: string[];
    /**
     * @generated from protobuf field: bool disabled = 4;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.MinMaxSetting
 */
export interface MinMaxSetting {
    /**
     * @generated from protobuf field: int32 min = 1;
     */
    min: number;
    /**
     * @generated from protobuf field: int32 max = 2;
     */
    max: number;
    /**
     * @generated from protobuf field: int32 step = 3;
     */
    step: number;
    /**
     * @generated from protobuf field: bool thousand_separator = 4;
     */
    thousandSeparator: boolean;
    /**
     * @generated from protobuf field: pistonpanel.v1.MinMaxSetting.Entry minEntry = 5;
     */
    minEntry?: MinMaxSetting_Entry;
    /**
     * @generated from protobuf field: pistonpanel.v1.MinMaxSetting.Entry maxEntry = 6;
     */
    maxEntry?: MinMaxSetting_Entry;
    /**
     * @generated from protobuf field: bool disabled = 7;
     */
    disabled: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.MinMaxSetting.Entry
 */
export interface MinMaxSetting_Entry {
    /**
     * @generated from protobuf field: string ui_name = 1;
     */
    uiName: string;
    /**
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: int32 def = 3;
     */
    def: number;
    /**
     * @generated from protobuf field: string placeholder = 4;
     */
    placeholder: string;
}
/**
 * A entry in the settings page
 *
 * @generated from protobuf message pistonpanel.v1.SettingEntry
 */
export interface SettingEntry {
    /**
     * Basically we only send a kv map to the server with every setting entry
     *
     * @generated from protobuf field: string key = 1;
     */
    key: string;
    /**
     * @generated from protobuf oneof: value
     */
    value: {
        oneofKind: "string";
        /**
         * @generated from protobuf field: pistonpanel.v1.StringSetting string = 2;
         */
        string: StringSetting;
    } | {
        oneofKind: "int";
        /**
         * @generated from protobuf field: pistonpanel.v1.IntSetting int = 3;
         */
        int: IntSetting;
    } | {
        oneofKind: "double";
        /**
         * @generated from protobuf field: pistonpanel.v1.DoubleSetting double = 4;
         */
        double: DoubleSetting;
    } | {
        oneofKind: "bool";
        /**
         * @generated from protobuf field: pistonpanel.v1.BoolSetting bool = 5;
         */
        bool: BoolSetting;
    } | {
        oneofKind: "combo";
        /**
         * @generated from protobuf field: pistonpanel.v1.ComboSetting combo = 6;
         */
        combo: ComboSetting;
    } | {
        oneofKind: "stringList";
        /**
         * @generated from protobuf field: pistonpanel.v1.StringListSetting string_list = 7;
         */
        stringList: StringListSetting;
    } | {
        oneofKind: "minMax";
        /**
         * @generated from protobuf field: pistonpanel.v1.MinMaxSetting min_max = 8;
         */
        minMax: MinMaxSetting;
    } | {
        oneofKind: undefined;
    };
}
// ...
export const StringSetting = new StringSetting$Type();
// ...
export const IntSetting = new IntSetting$Type();
// @generated message type with reflection information, may provide speed optimized methods
class DoubleSetting$Type extends MessageType<DoubleSetting> {
    constructor() {
        super("pistonpanel.v1.DoubleSetting", [
            { no: 1, name: "ui_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "description", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "def", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 4, name: "min", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 5, name: "max", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 6, name: "step", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 7, name: "placeholder", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 8, name: "thousand_separator", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 9, name: "decimal_scale", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 10, name: "fixed_decimal_scale", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 11, name: "disabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.DoubleSetting
 */
export const DoubleSetting = new DoubleSetting$Type();
// @generated message type with reflection information, may provide speed optimized methods
class BoolSetting$Type extends MessageType<BoolSetting> {
    constructor() {
        super("pistonpanel.v1.BoolSetting", [
            { no: 1, name: "ui_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "description", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "def", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "disabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.BoolSetting
 */
export const BoolSetting = new BoolSetting$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ComboSetting$Type extends MessageType<ComboSetting> {
    constructor() {
        super("pistonpanel.v1.ComboSetting", [
            { no: 1, name: "ui_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "description", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "options", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => ComboSetting_Option },
            { no: 4, name: "def", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "disabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.ComboSetting
 */
export const ComboSetting = new ComboSetting$Type();
// ...
export const ComboSetting_Option = new ComboSetting_Option$Type();
// ...
export const StringListSetting = new StringListSetting$Type();
// @generated message type with reflection information, may provide speed optimized methods
class MinMaxSetting$Type extends MessageType<MinMaxSetting> {
    constructor() {
        super("pistonpanel.v1.MinMaxSetting", [
            { no: 1, name: "min", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "max", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "step", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "thousand_separator", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "minEntry", kind: "message", T: () => MinMaxSetting_Entry },
            { no: 6, name: "maxEntry", kind: "message", T: () => MinMaxSetting_Entry },
            { no: 7, name: "disabled", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.MinMaxSetting
 */
export const MinMaxSetting = new MinMaxSetting$Type();
// ...
export const MinMaxSetting_Entry = new MinMaxSetting_Entry$Type();
// ...
export const SettingEntry = new SettingEntry$Type();